Phased changes for eventual migration to parquet
This commit will break the development branch of the package. But, as with all
things something must be broken before it can be fixed.

Wheel of time references aside, I am currently attempting to move HiCBricks
away from HDF and towards parquet. This will fulfill the initial vision of
bringing the package firmly into the apache ecosystem and thus all associated
methods.

Moreover, it fulfills the goal of the package not being a storage-layer but an
analytics layer.

By this commit

- Attributes will now be a part of the json file as opposed to the hdf file.
- Create_many_bricks is not creating any hdf files at the moment.
- Associated ranges have been moved from the HDF layer to an associated rds
file that is tracked via the json
    + Unlinking and removing the ranges will be possible but is not yet
    implemented.
- Chrominfo is being stored and tracked via the json
- A separate resolution specific segment has been implemented in the json
- The Container itself now contains header, resolutions and matrix info.